ai_phyton/requirements.txt
transformers==4.35.0
accelerate==0.24.1
bitsandbytes==0.41.3
torchvision
sentencepiece
fastapi
uvicorn
scipy